+++ src/main/java/dev/shadowsoffire/apotheosis/adventure/AdventureEvents.java
@@ -16,14 +16,15 @@ import dev.shadowsoffire.apotheosis.adventure.loot.LootController;
 import dev.shadowsoffire.apotheosis.util.Events;
 import dev.shadowsoffire.apotheosis.util.ItemAccess;
 import dev.shadowsoffire.attributeslib.api.ItemAttributeModifierEvent;
-import dev.shadowsoffire.attributeslib.api.ProjectileImpactCallback;
 import dev.shadowsoffire.placebo.events.AnvilLandCallback;
 import dev.shadowsoffire.placebo.events.GetEnchantmentLevelEvent;
 import dev.shadowsoffire.placebo.events.ItemUseEvent;
 import dev.shadowsoffire.placebo.reload.WeightedDynamicRegistry.IDimensional;
 import io.github.fabricators_of_create.porting_lib.entity.events.EntityEvents;
-import io.github.fabricators_of_create.porting_lib.entity.events.LivingEntityEvents;
-import io.github.fabricators_of_create.porting_lib.entity.events.PlayerEvents;
+import io.github.fabricators_of_create.porting_lib.entity.events.living.LivingEntityDamageEvents;
+import io.github.fabricators_of_create.porting_lib.entity.events.living.LivingEntityEvents;
+import io.github.fabricators_of_create.porting_lib.entity.events.living.LivingEntityLootEvents;
+import io.github.fabricators_of_create.porting_lib.entity.events.player.PlayerEvents;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import net.fabricmc.fabric.api.entity.FakePlayer;
 import net.fabricmc.fabric.api.event.player.PlayerBlockBreakEvents;
@@ -121,11 +122,10 @@ public class AdventureEvents {
     }
 
     public static void preventBossSuffocate() {
-        LivingEntityEvents.HURT.register((source, damaged, amount) -> {
-            if (source.is(DamageTypes.IN_WALL) && damaged.getCustomData().contains("apoth.boss")) {
-                return 0;
+        LivingEntityDamageEvents.HURT.register(e -> {
+            if (e.damageSource.is(DamageTypes.IN_WALL) && e.damaged.getCustomData().contains("apoth.boss")) {
+                e.setCanceled(true);
             }
-            return amount;
         });
 
     }
@@ -161,7 +161,7 @@ public class AdventureEvents {
      * This event handler allows affixes to react to arrows hitting something.
      */
     public static void impact() {
-        ProjectileImpactCallback.PROJECTILE_IMPACT.register((projectile, hitResult) -> {
+        EntityEvents.PROJECTILE_IMPACT.register((projectile, hitResult) -> {
             if (projectile instanceof AbstractArrow arrow) {
                 var affixes = AffixHelper.getAffixes(arrow);
                 affixes.values().forEach(inst -> inst.onArrowImpact(arrow, hitResult, hitResult.getType()));
@@ -172,19 +172,22 @@ public class AdventureEvents {
     }
 
     public static void onDamage() {
-        LivingEntityEvents.HURT.register((src, ent, amount) -> {
-            Adventure.Affixes.MAGICAL.getOptional().ifPresent(afx -> afx.onHurt(src));
+        LivingEntityDamageEvents.HURT.register(e -> {
+            Adventure.Affixes.MAGICAL.getOptional().ifPresent(afx -> afx.onHurt(e));
+            DamageSource src = e.damageSource;
+            LivingEntity ent = e.damaged;
+            float amount = e.damageAmount;
             for (ItemStack s : ent.getAllSlots()) {
                 var affixes = AffixHelper.getAffixes(s);
                 for (AffixInstance inst : affixes.values()) {
                     amount = inst.onHurt(src, ent, amount);
                 }
             }
-            return amount;
+            e.damageAmount = amount;
         });
     }
 
-    public static void onItemUse() {
+    public static void onItemUse() { //TODO Item use Event from placebo
   /*      ItemUseEvent
             ItemStack s = player.getItemInHand(hand);
             var affixes = AffixHelper.getAffixes(s);
@@ -225,7 +228,7 @@ public class AdventureEvents {
     }
 
     public static void dropsHigh() {
-        LivingEntityEvents.DROPS.register((target, source, drops, lootingLevel, recentlyHit) -> {
+        LivingEntityLootEvents.DROPS.register((target, source, drops, lootingLevel, recentlyHit) -> {
             if (source.getEntity() instanceof ServerPlayer p && target instanceof Monster) {
                 if (p instanceof FakePlayer) return false;
                 float chance = AdventureConfig.gemDropChance + (target.getCustomData().contains("apoth.boss") ? AdventureConfig.gemBossBonus : 0);
@@ -240,7 +243,7 @@ public class AdventureEvents {
     }
 
     public static void drops() {
-        LivingEntityEvents.DROPS.register((target, source, drops, lootingLevel, recentlyHit) -> {
+        LivingEntityLootEvents.DROPS.register((target, source, drops, lootingLevel, recentlyHit) -> {
             AdventureModule.LOGGER.info("present: {}", Adventure.Affixes.FESTIVE.getOptional().isPresent());
             Adventure.Affixes.FESTIVE.getOptional().ifPresent(afx -> afx.drops(target, source, drops));
             TelepathicAffix.drops(source, drops);
@@ -268,8 +271,10 @@ public class AdventureEvents {
     }
 
     public static void speed() {
-        PlayerEvents.BREAK_SPEED.register(e -> {
-            Adventure.Affixes.OMNETIC.getOptional().ifPresent(afx -> afx.speed(e));
+        PlayerEvents.BREAK_SPEED.register((player, state, pos, speed) -> {
+            AtomicReference<Float> finalSpeed = new AtomicReference<>(speed);
+            Adventure.Affixes.OMNETIC.getOptional().ifPresent(afx -> finalSpeed.set(afx.speed(player, state, pos, speed)));
+            return finalSpeed.get();
         });
 
     }
@@ -281,19 +286,19 @@ public class AdventureEvents {
     }
 
     public static void special() {
-        LivingEntityEvents.CHECK_SPAWN.register((mob, level, x, y, z, spawner, spawnReason) -> {
+        LivingEntityEvents.NATURAL_SPAWN.register((mob, x, y, z, level, spawner, type) -> {
             if (level.getRandom().nextFloat() <= AdventureConfig.randomAffixItem && mob instanceof Monster) {
                 Player player = level.getNearestPlayer(x, y, z, -1, false);
-                if (player == null) return false;
+                if (player == null) return TriState.DEFAULT;
                 ItemStack affixItem = LootController.createRandomLootItem(level.getRandom(), null, player, (ServerLevel) mob.level());
-                if (affixItem.isEmpty()) return false;
+                if (affixItem.isEmpty()) return TriState.DEFAULT;
                 affixItem.getOrCreateTag().putBoolean("apoth_rspawn", true);
                 LootCategory cat = LootCategory.forItem(affixItem);
                 EquipmentSlot slot = cat.getSlots()[0];
                 mob.setItemSlot(slot, affixItem);
                 mob.setGuaranteedDrop(slot);
             }
-            return false;
+            return TriState.DEFAULT;
         });
 
     }