+++ changelog.md
@@ -1,4 +1,5 @@
-
+## TODO
+* Complete port of Spawner module and Adventure module
 
 ## 0.0.1
 * Initial release